requirements.txt
django
djangorestframework
django-cors-headers
Pillow
drf-yasg
django-modeltranslation
dj-rest-auth
django-allauth
djangorestframework-simplejwt
django-filter